Desktop/this-is_i/DISCORD/Creating a Discord Bot with AI and RAG Features/requirements.txt
discord.py>=2.3.0
requests>=2.28.0
python-dotenv>=0.19.0
beautifulsoup4>=4.11.0
feedparser>=6.0.0
ollama>=0.1.0
sqlite3
asyncio
pathlib
datetime
json
pillow>=9.0.0
aiohttp>=3.8.0
python-nmap>=0.7.0
psutil>=5.9.0
matplotlib>=3.5.0
numpy>=1.21.0